requirements/base.txt
Django==1.8
django-auth-ldap==1.2.0
django-crispy-forms==1.4.0
django-multiselectfield==0.1.2
django-report-builder==3.2.0
django-report-utils==0.3.15
django-reversion==1.8.5
MySQL-python==1.2.5
python-dateutil==2.2
requests==2.4.0
xlwt==1.3.0
